+++ src/main/java/com/puppycrawl/tools/checkstyle/checks/indentation/LineSet.java
@@ -52,7 +52,7 @@ public class LineSet {
      */
     public int firstLineCol() {
         final Object firstLineKey = lines.firstKey();
-        return lines.get(firstLineKey).intValue();
+        return lines.get(firstLineKey);
     }
 
     /**
@@ -61,7 +61,7 @@ public class LineSet {
      * @return the line number of the first line
      */
     public int firstLine() {
-        return lines.firstKey().intValue();
+        return lines.firstKey();
     }
 
     /**
@@ -70,7 +70,7 @@ public class LineSet {
      * @return the line number of the last line
      */
     public int lastLine() {
-        return lines.lastKey().intValue();
+        return lines.lastKey();
     }
 
     /**
